small code editor refactoring
